pyproject.toml
[project]
name = "gfeed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Dhruvin Shah", email = "dhruvin3@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "pydantic>=2.10.3",
    "pyopml>=1.0.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = ["pyinstaller>=6.11.1"]

[project.scripts]
gfeed = "gfeed.__main__:cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
